% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\alias{random}
\alias{runthreads}
\title{Hello Rust!}
\usage{
hello()

random()

runthreads()
}
\value{
a value generated in Rust (a string, random number, and NULL respectively).
}
\description{
Examples of rust functions via C.
}
\examples{
hello()
random()
runthreads()
}
